+++ scripts/calculateRevenue/protocols/aave/index.ts
@@ -8,6 +8,7 @@ import { getATokenScaledBalances } from './aToken'
 import { getATokenScaledBalanceHistory } from './subgraph'
 import { RAY, rayDiv, rayMul } from './math'
 import { getUSDPrices } from './oracle'
+import { calculateOverlap, createSegments } from './utils'
 
 interface ReserveFactorSegment {
   value: bigint
@@ -358,30 +359,4 @@ function totalRevenueInUSD(
   }
 
   return totalRevenueInUSD
-}
-
-function calculateOverlap(
-  start1: number,
-  end1: number,
-  start2: number,
-  end2: number,
-): number {
-  const overlapStart = Math.max(start1, start2)
-  const overlapEnd = Math.min(end1, end2)
-  return Math.max(0, overlapEnd - overlapStart)
-}
-
-function createSegments<T, R>(
-  snapshots: T[],
-  transformFn: (current: T, next: T) => R,
-): R[] {
-  const segments: R[] = []
-
-  for (let i = 0; i < snapshots.length - 1; i++) {
-    const current = snapshots[i]
-    const next = snapshots[i + 1]
-    segments.push(transformFn(current, next))
-  }
-
-  return segments
 }
